move map_properties
