+++ napari/layers/shapes/shapes.py
@@ -16,14 +16,14 @@ from ...utils.colormaps.standardize_color import (
 from ...utils.events import Event
 from ...utils.misc import ensure_iterable
 from ..base import Layer
-from ..utils.color_manager_utils import guess_continuous
+from ..utils.color_manager_utils import guess_continuous, map_property
 from ..utils.color_transformations import (
     ColorType,
     normalize_and_broadcast_colors,
     transform_color_cycle,
     transform_color_with_defaults,
 )
-from ..utils.layer_utils import dataframe_to_properties, map_property
+from ..utils.layer_utils import dataframe_to_properties
 from ..utils.text import TextManager
 from ._shape_list import ShapeList
 from ._shapes_constants import (
